Add toHexString to public api
Moves `toHexString` from __test__ to index.ts, making it a public
convenience function. This was motivated by a need for a way to convert
hashes into hex string for hyperledger/sawtooth-sabre compatibility.
